Move isLatestReleaseBranch to BranchLogic
