+++ src/main/ts/core/BranchLogic.ts
@@ -5,8 +5,10 @@ import * as PromiseUtils from '../utils/PromiseUtils';
 import { showStringOrUndefined } from '../utils/StringUtils';
 import * as Git from '../utils/Git';
 import * as Version from './Version';
+import { compareMajorMinorVersions } from './Version';
 import * as PreRelease from './PreRelease';
 import * as PackageJson from './PackageJson';
+import * as ArrayUtils from '../utils/ArrayUtils';
 
 type MajorMinorVersion = Version.MajorMinorVersion;
 type Version = Version.Version;
@@ -163,4 +165,13 @@ export const getBranchDetails = async (dir: string): Promise<BranchDetails> => {
       };
     }
   }
+};
+
+export const isLatestReleaseBranch = async (getBranches: () => Promise<string[]>, branchName: string): Promise<boolean> => {
+  const branches = await getBranches();
+  const versions = await PromiseUtils.filterMap(branches, versionFromReleaseBranch);
+  const greatestOpt = ArrayUtils.greatest(versions, compareMajorMinorVersions);
+  const greatest = await PromiseUtils.optionToPromise(greatestOpt, 'Could not find any release branches with valid names.');
+  const releaseBranchName = getReleaseBranchName(greatest);
+  return branchName === releaseBranchName;
 };
